+++ packages/primereact/src/RadioWidget/RadioWidget.tsx
@@ -1,7 +1,8 @@
 import {
   ariaDescribedByIds,
+  enumOptionValueDecoder,
+  enumOptionValueEncoder,
   enumOptionsIsSelected,
-  enumOptionsValueForIndex,
   optionId,
   FormContextType,
   RJSFSchema,
@@ -21,9 +22,10 @@ export default function RadioWidget<T = any, S extends StrictRJSFSchema = RJSFSc
   const { id, htmlName, value, disabled, readonly, onChange, onBlur, onFocus, options } = props;
   const primeProps = (options.prime || {}) as object;
   const { enumOptions, enumDisabled, emptyValue } = options;
+  const useRealValues = !!options.useRealOptionValues;
 
   const _onChange = (e: RadioButtonChangeEvent) => {
-    onChange(enumOptionsValueForIndex<S>(e.value, enumOptions, emptyValue));
+    onChange(enumOptionValueDecoder<S>(e.value, enumOptions, useRealValues, emptyValue));
   };
 
   const _onBlur = () => onBlur(id, value);
@@ -44,7 +46,7 @@ export default function RadioWidget<T = any, S extends StrictRJSFSchema = RJSFSc
                 onFocus={_onFocus}
                 onBlur={_onBlur}
                 onChange={_onChange}
-                value={String(index)}
+                value={enumOptionValueEncoder(option.value, index, useRealValues)}
                 checked={checked}
                 disabled={disabled || itemDisabled || readonly}
                 aria-describedby={ariaDescribedByIds(id)}