+++ packages/primereact/src/SelectWidget/SelectWidget.tsx
@@ -2,8 +2,9 @@ import { FocusEvent } from 'react';
 import { Dropdown } from 'primereact/dropdown';
 import {
   ariaDescribedByIds,
-  enumOptionsIndexForValue,
-  enumOptionsValueForIndex,
+  enumOptionSelectedValue,
+  enumOptionValueDecoder,
+  enumOptionValueEncoder,
   FormContextType,
   RJSFSchema,
   StrictRJSFSchema,
@@ -50,19 +51,19 @@ function SingleSelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F 
   ...dropdownProps
 }: WidgetProps<T, S, F>) {
   const { enumOptions, enumDisabled, emptyValue: optEmptyVal } = options;
+  const useRealValues = !!options.useRealOptionValues;
   const primeProps = (options.prime || {}) as object;
 
   multiple = typeof multiple === 'undefined' ? false : multiple;
 
   const emptyValue = multiple ? [] : '';
-  const isEmpty = typeof value === 'undefined' || (multiple && value.length < 1) || (!multiple && value === emptyValue);
 
-  const _onChange = (e: { value: any }) => onChange(enumOptionsValueForIndex<S>(e.value, enumOptions, optEmptyVal));
+  const _onChange = (e: { value: any }) =>
+    onChange(enumOptionValueDecoder<S>(e.value, enumOptions, useRealValues, optEmptyVal));
   const _onBlur = ({ target }: FocusEvent<HTMLInputElement>) =>
-    onBlur(id, enumOptionsValueForIndex<S>(target && target.value, enumOptions, optEmptyVal));
+    onBlur(id, enumOptionValueDecoder<S>(target && target.value, enumOptions, useRealValues, optEmptyVal));
   const _onFocus = ({ target }: FocusEvent<HTMLInputElement>) =>
-    onFocus(id, enumOptionsValueForIndex<S>(target && target.value, enumOptions, optEmptyVal));
-  const selectedIndexes = enumOptionsIndexForValue<S>(value, enumOptions, multiple);
+    onFocus(id, enumOptionValueDecoder<S>(target && target.value, enumOptions, useRealValues, optEmptyVal));
   const { ...dropdownRemainingProps } = dropdownProps;
 
   return (
@@ -70,10 +71,10 @@ function SingleSelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F 
       id={id}
       name={htmlName || id}
       {...primeProps}
-      value={!isEmpty && typeof selectedIndexes !== 'undefined' ? selectedIndexes : emptyValue}
+      value={enumOptionSelectedValue<S>(value, enumOptions, !!multiple, useRealValues, emptyValue)}
       options={(enumOptions ?? []).map(({ value, label }, i: number) => ({
         label,
-        value: String(i),
+        value: enumOptionValueEncoder(value, i, useRealValues),
         disabled: Array.isArray(enumDisabled) && enumDisabled.indexOf(value) !== -1,
       }))}
       onChange={_onChange}
@@ -103,27 +104,27 @@ function MultiSelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F e
   onFocus,
 }: WidgetProps<T, S, F>) {
   const { enumOptions, enumDisabled, emptyValue: optEmptyVal } = options;
+  const useRealValues = !!options.useRealOptionValues;
   const primeProps = (options.prime || {}) as object;
 
   const emptyValue = multiple ? [] : '';
-  const isEmpty = typeof value === 'undefined' || (multiple && value.length < 1) || (!multiple && value === emptyValue);
 
-  const _onChange = (e: { value: any }) => onChange(enumOptionsValueForIndex<S>(e.value, enumOptions, optEmptyVal));
+  const _onChange = (e: { value: any }) =>
+    onChange(enumOptionValueDecoder<S>(e.value, enumOptions, useRealValues, optEmptyVal));
   const _onBlur = ({ target }: FocusEvent<HTMLInputElement>) =>
-    onBlur(id, enumOptionsValueForIndex<S>(target && target.value, enumOptions, optEmptyVal));
+    onBlur(id, enumOptionValueDecoder<S>(target && target.value, enumOptions, useRealValues, optEmptyVal));
   const _onFocus = ({ target }: FocusEvent<HTMLInputElement>) =>
-    onFocus(id, enumOptionsValueForIndex<S>(target && target.value, enumOptions, optEmptyVal));
-  const selectedIndexes = enumOptionsIndexForValue<S>(value, enumOptions, multiple);
+    onFocus(id, enumOptionValueDecoder<S>(target && target.value, enumOptions, useRealValues, optEmptyVal));
 
   return (
     <MultiSelect
       id={id}
       name={htmlName || id}
       {...primeProps}
-      value={!isEmpty && typeof selectedIndexes !== 'undefined' ? selectedIndexes : emptyValue}
+      value={enumOptionSelectedValue<S>(value, enumOptions, multiple, useRealValues, emptyValue)}
       options={(enumOptions ?? []).map(({ value, label }, i: number) => ({
         label,
-        value: String(i),
+        value: enumOptionValueEncoder(value, i, useRealValues),
         disabled: Array.isArray(enumDisabled) && enumDisabled.indexOf(value) !== -1,
       }))}
       onChange={_onChange}