+++ packages/react-bootstrap/src/SelectWidget/SelectWidget.tsx
@@ -2,9 +2,10 @@ import { ChangeEvent, FocusEvent } from 'react';
 import FormSelect from 'react-bootstrap/FormSelect';
 import {
   ariaDescribedByIds,
+  enumOptionSelectedValue,
+  enumOptionValueDecoder,
+  enumOptionValueEncoder,
   FormContextType,
-  enumOptionsIndexForValue,
-  enumOptionsValueForIndex,
   RJSFSchema,
   StrictRJSFSchema,
   WidgetProps,
@@ -34,6 +35,7 @@ export default function SelectWidget<
   const { enumOptions, enumDisabled, emptyValue: optEmptyValue } = options;
 
   const emptyValue = multiple ? [] : '';
+  const useRealValues = !!options.useRealOptionValues;
 
   function getValue(event: FocusEvent | ChangeEvent | any, multiple?: boolean) {
     if (multiple) {
@@ -45,14 +47,14 @@ export default function SelectWidget<
       return event.target.value;
     }
   }
-  const selectedIndexes = enumOptionsIndexForValue<S>(value, enumOptions, multiple);
+  const selectValue = enumOptionSelectedValue<S>(value, enumOptions, !!multiple, useRealValues, emptyValue);
   const showPlaceholderOption = !multiple && schema.default === undefined;
 
   return (
     <FormSelect
       id={id}
       name={htmlName || id}
-      value={typeof selectedIndexes === 'undefined' ? emptyValue : selectedIndexes}
+      value={selectValue}
       required={required}
       multiple={multiple}
       disabled={disabled || readonly}
@@ -62,19 +64,19 @@ export default function SelectWidget<
         onBlur &&
         ((event: FocusEvent) => {
           const newValue = getValue(event, multiple);
-          onBlur(id, enumOptionsValueForIndex<S>(newValue, enumOptions, optEmptyValue));
+          onBlur(id, enumOptionValueDecoder<S>(newValue, enumOptions, useRealValues, optEmptyValue));
         })
       }
       onFocus={
         onFocus &&
         ((event: FocusEvent) => {
           const newValue = getValue(event, multiple);
-          onFocus(id, enumOptionsValueForIndex<S>(newValue, enumOptions, optEmptyValue));
+          onFocus(id, enumOptionValueDecoder<S>(newValue, enumOptions, useRealValues, optEmptyValue));
         })
       }
       onChange={(event: ChangeEvent) => {
         const newValue = getValue(event, multiple);
-        onChange(enumOptionsValueForIndex<S>(newValue, enumOptions, optEmptyValue));
+        onChange(enumOptionValueDecoder<S>(newValue, enumOptions, useRealValues, optEmptyValue));
       }}
       aria-describedby={ariaDescribedByIds(id)}
     >
@@ -82,7 +84,7 @@ export default function SelectWidget<
       {(enumOptions as any).map(({ value, label }: any, i: number) => {
         const disabled: any = Array.isArray(enumDisabled) && (enumDisabled as any).indexOf(value) != -1;
         return (
-          <option key={i} id={label} value={String(i)} disabled={disabled}>
+          <option key={i} id={label} value={enumOptionValueEncoder(value, i, useRealValues)} disabled={disabled}>
             {label}
           </option>
         );