+++ packages/shadcn/src/SelectWidget/SelectWidget.tsx
@@ -1,7 +1,8 @@
 import {
   ariaDescribedByIds,
-  enumOptionsIndexForValue,
-  enumOptionsValueForIndex,
+  enumOptionSelectedValue,
+  enumOptionValueDecoder,
+  enumOptionValueEncoder,
   FormContextType,
   RJSFSchema,
   StrictRJSFSchema,
@@ -33,23 +34,23 @@ export default function SelectWidget<
   onChange,
   onBlur,
   onFocus,
-  defaultValue,
   placeholder,
   rawErrors = [],
   className,
 }: WidgetProps<T, S, F>) {
   const { enumOptions, enumDisabled, emptyValue: optEmptyValue } = options;
+  const useRealValues = !!options.useRealOptionValues;
 
   const _onFancyFocus = () => {
-    onFocus(id, enumOptionsValueForIndex<S>(value, enumOptions, optEmptyValue));
+    onFocus(id, enumOptionValueDecoder<S>(value, enumOptions, useRealValues, optEmptyValue));
   };
 
   const _onFancyBlur = () => {
-    onBlur(id, enumOptionsValueForIndex<S>(value, enumOptions, optEmptyValue));
+    onBlur(id, enumOptionValueDecoder<S>(value, enumOptions, useRealValues, optEmptyValue));
   };
 
   const items = (enumOptions as any)?.map(({ value, label }: any, index: number) => ({
-    value: multiple ? value : index.toString(),
+    value: multiple ? value : enumOptionValueEncoder(value, index, useRealValues),
     label: label,
     index,
     disabled: Array.isArray(enumDisabled) && enumDisabled.includes(value),
@@ -62,9 +63,9 @@ export default function SelectWidget<
       {!multiple ? (
         <FancySelect
           items={items}
-          selected={enumOptionsIndexForValue<S>(value ?? defaultValue, enumOptions, false) as unknown as string}
+          selected={enumOptionSelectedValue<S>(value, enumOptions, false, useRealValues, '') as string}
           onValueChange={(selectedValue) => {
-            onChange(enumOptionsValueForIndex<S>(selectedValue, enumOptions, optEmptyValue));
+            onChange(enumOptionValueDecoder<S>(selectedValue, enumOptions, useRealValues, optEmptyValue));
           }}
           autoFocus={autofocus}
           disabled={disabled || readonly}
@@ -85,7 +86,7 @@ export default function SelectWidget<
           items={items}
           selected={value}
           onValueChange={(values) => {
-            onChange(enumOptionsValueForIndex<S>(values, enumOptions, optEmptyValue));
+            onChange(enumOptionValueDecoder<S>(values.map(String), enumOptions, useRealValues, optEmptyValue));
           }}
           onFocus={_onFancyFocus}
           onBlur={_onFancyBlur}